samples-v2/microsoft/python/quickstart/requirements.txt
azure-ai-projects>=2.0.0a20250915020
azure-identity
python-dotenv
